Fix pygen runtime: use int() instead of long() for Python 3 compatibility
long() was removed in Python 3 (PEP 237). Python 3's int is already
arbitrary-precision, and Python 2's int() auto-promotes to long when
needed, so int() is correct for both versions.

+++ PySupport/py_support.py
@@ -143,7 +143,7 @@ def _crashExp(x):
     raise x#(ValueError x)
 
 def _bigIntOfString(s):
-    return long(s)
+    return int(s)
 
 def _numberOfString(s):
     return float(s)
